[Core] Factor out input preprocessing to a separate class (#7329)
